requirements.txt
chromadb==0.5.23
langchain==0.3.26
sentence-transformers==3.3.1
langgraph==0.2.54
langchain-openai==0.2.14
openai==1.57.0
requests==2.32.3
python-dotenv==1.0.1
serpapi==0.1.5
loguru==0.7.2
trafilatura==1.12.2
lxml_html_clean==0.3.1
streamlit==1.40.1
tiktoken==0.8.0
scikit-learn==1.5.2